Add session bind
