+++ src/lib.rs
@@ -1,3 +1,4 @@
 pub mod encoding;
 pub mod msg;
+pub mod session_bind;
 pub mod ssh_agent;
